back/pyproject.toml
[project]
name = "monitoringbattery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.17.1",
    "annotated-types>=0.7.0",
    "anyio>=4.11.0",
    "argon2-cffi>=25.1.0",
    "argon2-cffi-bindings>=25.1.0",
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "bleach>=6.3.0",
    "cffi>=2.0.0",
    "click>=8.3.0",
    "colorama>=0.4.6",
    "cryptography>=46.0.3",
    "dnspython>=2.8.0",
    "ecdsa>=0.19.1",
    "email-validator>=2.3.0",
    "fastapi>=0.120.4",
    "greenlet>=3.2.4",
    "h11>=0.16.0",
    "httptools>=0.7.1",
    "idna>=3.11",
    "mako>=1.3.10",
    "markupsafe>=3.0.3",
    "passlib>=1.7.4",
    "psycopg>=3.2.12",
    "psycopg-binary>=3.2.12",
    "pyasn1>=0.6.1",
    "pycparser>=2.23",
    "pydantic>=2.12.3",
    "pydantic-core>=2.41.4",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "python-slugify>=8.0.4",
    "pyyaml>=6.0.3",
    "rsa>=4.9.1",
    "six>=1.17.0",
    "sniffio>=1.3.1",
    "sqlalchemy>=2.0.44",
    "starlette>=0.49.3",
    "text-unidecode>=1.3",
    "typing-extensions>=4.15.0",
    "typing-inspection>=0.4.2",
    "tzdata>=2025.2",
    "uvicorn[standard]>=0.38.0",
    "watchfiles>=1.1.1",
    "webencodings>=0.5.1",
    "websockets>=15.0.1",
]
